Guard the timer step with a lock.
When a message is shown and the user clicks the systray icon, it generates
two events calling the slotStep() function.
To workaround this, introduce a lock lasting 100 ms. to prevent the other
event from calling mTimer->step().

I also tried this with a 0 ms. timeout, but apparently when the user
clicks, it goes lock -> step -> release -> lock -> release -> step.
Apparently the second event comes in too late, and the event loop already
has kicked in and released the lock halfway.