Confirmation dialogs were shown multiple times on multiple clicks.
The semantics of the step lock have been reversed: 0 means released and any
value >0 means locked. The lock is active 100 ms after the first click, and
while showing a confirmation dialog (which usually surpasses the 100 ms).
This is to prevent a confirmation dialog to be shown multiple times when a user
clicks, waits >100 ms and then clicks again.